Refactor DoubleLinkedList class to include end pointer
The DoubleLinkedList class has been refactored to include an end
pointer. Previously, the class only had a start pointer to track the
beginning of the linked list. Adding the end pointer allows for
efficient insertion at the end and deletion of the last element. This
change improves the overall functionality and performance of the
DoubleLinkedList data structure.

Fixes: #123